requirements.txt
customtkinter
opencv-python
scikit-learn
numpy
rembg
Pillow
requests
SpeechRecognition
pyaudio
flask
qrcode